level_three/resume-improver/backend/requirements.txt
Django
djangorestframework
openai
python-dotenv